Improve some stuff, add docs and remove unused stuff (connectv1 annoying bits).
